requirements.txt
# Core Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23

# Face Recognition
insightface==0.7.3
onnxruntime==1.16.3

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
humanize==4.9.0

# Development (Optional)
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0